% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{predict_solar}
\alias{predict_solar}
\title{Perform a ARIMA model on a specific column}
\usage{
predict_solar(SolarEnergy_data, col_of_interest)
}
\arguments{
\item{SolarEnergy_data}{A data frame containing solar data.}

\item{col_of_interest}{A string corresponding to the column of the `SolarEnergy` data frame that you wish to perform forecasting on.}
}
\value{
A plot
}
\description{
Perform a ARIMA model on a specific column
}
\examples{
# Forecast the `Temperature`
predict_solar(SolarEnergy,
              col_of_interest = "Temperature")
}
